// static/netplay/bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const Input = __t.object("Input", {
  guest: __t.identity(),
  host: __t.identity(),
  seq: __t.u32(),
  bits: __t.u32(),
  at: __t.timestamp(),
});
export type Input = __Infer<typeof Input>;

export const ReapTimer = __t.object("ReapTimer", {
  scheduledId: __t.u64(),
  scheduledAt: __t.scheduleAt(),
});
export type ReapTimer = __Infer<typeof ReapTimer>;

export const Session = __t.object("Session", {
  host: __t.identity(),
  gameId: __t.string(),
  title: __t.string(),
  stage: __t.u32(),
  score: __t.u64(),
  players: __t.u32(),
  joinable: __t.bool(),
  guest: __t.option(__t.identity()),
  startedAt: __t.timestamp(),
  updatedAt: __t.timestamp(),
});
export type Session = __Infer<typeof Session>;

export const Snapshot = __t.object("Snapshot", {
  host: __t.identity(),
  seq: __t.u32(),
  at: __t.timestamp(),
  blob: __t.string(),
});
export type Snapshot = __Infer<typeof Snapshot>;